refactor(form): move form utility location and update description
