+++ libs/core/src/utils/utils.ts
@@ -5,30 +5,3 @@ export function format(first: string, middle: string, last: string): string {
 export const hasShadowDom = (el: HTMLElement) => {
   return !!el.shadowRoot && !!(el as any).attachShadow
 }
-
-export const isRequired = (target, component) => {
-  if ( !target || !component ) return;
-  if (component.required === true) {
-    (target.checkValidity() === false) ? component.invalid = true : component.invalid = false;
-  }
-}
-
-/**
- * Create id for messaging
- */
-export const messageId = (id: string, messageType: string) => {
-  return `${id}__${messageType}-message`;
-};
-
-/**
- * Assign aria-description id to relate messages with form element
- */
-export const assignDescription = (id: string, invalid: boolean, helperMessage: string) => {
-  if (!helperMessage) return
-
-  let relatedId = messageId(id, 'helper')
-
-  if (invalid) relatedId = messageId(id, 'error');
-
-  return relatedId;
-};
